Further improve `.python-version` and `runtime.txt` error handling
Some further improvements on-top of those made in #1958, based on build
error message scenarios seen in Honeycomb.

Now, if `.python-version` contains a single ESC control code (which gets
categorised as "very short file"), or contains any of the ASCI control
codes that result in the file being categorised as "data" (such as NUL),
then the "invalid python version" error message variant is shown instead
of the "invalid text encoding" variant.

In addition, any NULs in the file are substituted with a placeholder to
avoid this Bash warning:

```
/tmp/buildpack/lib/python_version.sh: line 104: warning: command substitution: ignored null byte in input
```

See:
https://manpages.ubuntu.com/manpages/noble/en/man1/file.1.html

GUS-W-20220514.