Add more messaging to the build log recommending uv
The package manager uv offers many benefits for users over the
alternatives, such as it:
- Supporting lockfiles (unlike pip)
- Being much faster
- Giving more helpful error messages
- Being actively maintained by a full time team

As such we want to encourage users who may not have heard of it (or be
aware that the buildpack now supports it) to try it out - and so now
mention uv at the env of the build log for successful builds.

GUS-W-19124925.