Stop adding the current working directory to `PATH` (#1987)
Currently for historic reasons the buildpack adds the current working
directory to `PATH`, due to adding an empty colon-delimited segment
before the default Ubuntu `PATH`.

This can cause errors if the app has a script/binary in its root
directory that has the same name as a command the buildpack tries
to run. For example, on Honeycomb I've seen internal errors when
the buildpack attempts to locate the bundled pip, due to the `tail`
command being run not being the command we expected:

```
-----> Installing pip 25.2, setuptools 70.3.0 and wheel 0.45.1
./tail: line 1: heroku: command not found

 !     Internal Error: Unable to locate the Python stdlib's bundled pip.
```

Removing this customisation of `PATH` is a breaking change, but safer
in the long run (and also matches the Python CNB behaviour).

Any apps that call a script or binary that's in the root directory
of their app from `bin/pre_compile` or `bin/post_compile` hook
(or any other subprocess called by the buildpack) will now need to do
so via an explicit relative path. For example, use `./my_command`
not `my_command` when calling the script/binary.

This change doesn't affect:
- app behaviour at run-time (only during the build)
- apps that call scripts in a subdirectory (eg `foo/my_command`)
- apps that already use a relative path for scripts in the app's
  root directory (eg `./my_command`)

GUS-W-20472356.