Sunset implicit setup.py support (#1992)
Previously if an app had only a `setup.py` file and no Python package
manager files (such as `requirements.txt`, `Pipfile.lock`, `Poetry.lock`
or `uv.lock`), the buildpack would install the project using
`pip install --editable .`

However, this implicit fallback doesn't make sense now that the
buildpack supports multiple package managers, and so has to guess
which package manager to use, and whether to install the project in
editable mode or not.

As such, in #1897 this fallback was deprecated, and is now being sunset.
This also brings the classic Python buildpack's behaviour in line with
the Python CNB.

Apps that only have a `setup.py` file will now need to add an explicit
`requirements.txt` file containing:

```
--editable .
```

GUS-W-19275444.