Error when multiple package manager files are found (#1993)
Currently the buildpack will use whichever package manager it finds
first, if the files of multiple package managers are found.

This occasionally results in support tickets where the user believes the
build to not be installing dependencies correctly, when in fact they are
adding dependencies to the wrong package manager file.

It also means historic users of third-party Poetry and uv buildpacks
might not realise that the buildpack now natively supports those package
managers (in a much more efficient way), if they missed our release
announcements.

As such, in November 2024 in #1692 a warning was added, which we're now
converting to an error.

Users with multiple package manager files committed to their Git repo
will need to pick one package manager and delete the files relating to
the others.

And users who are still using the third-party Poetry or uv buildpacks
will need to remove those buildpacks from their app:
https://devcenter.heroku.com/articles/managing-buildpacks#remove-classic-buildpacks

Closes #1691.
GUS-W-18915632.