Update pip from 25.2 to 25.3
Changelog:
https://pip.pypa.io/en/stable/news/#v25-3

Additionally, we now no longer install the `wheel` package when using
pip with Python 3.12 and older, since:

- With modern setuptools, it's no longer necessary to manually install
  wheel, since setuptools will trigger its installation if needed.
- As of pip 25.3, pip prepares/builds dependencies inside an isolated
  ephemeral build environment (~venv) that means even if wheel is
  installed globally, it won't be used anyway.
- The Python CNB (and much of the rest of the Python ecosystem) doesn't
  install wheel globally any more either.

This doesn't affect apps using either a different package manager, or
that are using pip with Python 3.13 or newer, since we already didn't
install the wheel package in those environments.

GUS-W-17605676.
GUS-W-20804745.